tests/wpt/web-platform-tests/css/css-logical/parsing/inset-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Logical Properties and Values: parsing inset with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-logical/#propdef-inset">
<meta name="assert" content="inset supports only the grammar '<'top'>{1,4}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("inset", "none");
test_invalid_value("inset", "20%, calc(10px - 0.5em)");
test_invalid_value("inset", "10px auto 20px auto 30px");
</script>
</body>
</html>
